+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 _Compatible with: [`smart-contract`](https://github.com/ndidplatform/smart-contract) v0.13.0_
 
 BUG FIXES:
+
 - Fix check `chain_id` logic when recieve MQ
 
 ## 0.12.0 (November 21, 2018)