+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 _Compatible with: [`smart-contract`](https://github.com/ndidplatform/smart-contract) v.development_
 
+IMPROVEMENTS:
+
+- Reuse MQ connection when possible if there are multiple send message calls and the destination IP and port are the same.
+
 BUG FIXES:
 
 - Remove both chain ID and latest block height files first when handling new chain function is called to prevent invalid value in latest block height file in case the server stops before it is able to write a block height of a new chain.